Finished basic Roslyn + Legacy OmniSharpServer
Added in a prebuilt roslyn so that we can ship both
omnisharp-server (legacy), as well as omnisharp-roslyn(roslyn).
In order to actually use both inside of sublime, I added
a configuration option in `OmniSharpSublime.sublime-settings`
that allows users to change the active omnisharp server.
Changes in the server will be realized upon relaunching sublime
text. I am planning on adding in a feature to switch the active
server from a menu instead of editing the config file.